% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe_models.R
\docType{data}
\name{model_en}
\alias{model_en}
\title{UDPIPE english model}
\format{
An object of class \code{udpipe_model} of length 2.
}
\usage{
model_en
}
\description{
UDPIPE english model
}
\keyword{datasets}
