% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe_models.R
\docType{data}
\name{model_sp}
\alias{model_sp}
\title{UDPIPE spanish model}
\format{
An object of class \code{udpipe_model} of length 2.
}
\usage{
model_sp
}
\description{
UDPIPE spanish model
}
\keyword{datasets}
